set up react context for chatroom list and selected chatroom. fix up formatting in css
